requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
requests==2.31.0
playwright==1.40.0
python-dotenv==1.0.0
google-search-results==2.4.2
